restructed store to be more modulized
