+++ resources/js/Store/mock/exampleMotifs.js
