+++ resources/js/Store/mock/exampleProducts.js
