+++ resources/js/Store/modules/cart/actions.js
@@ -1,72 +1,6 @@
-import axios from 'axios'
-import Vue from 'vue'
-import Vuex from 'vuex'
+import {baseColor, baseItem, baseProduct, baseSize, baseMotif} from "../../mock/baseData"
 
-import {baseColor, baseItem, baseOption, baseProduct, baseSize, baseMotif} from "./Mock/baseData"
-import exampleProducts from "./Mock/exampleProducts"
-import exampleMotifs from "./Mock/exampleProducts"
-
-Vue.use(Vuex)
-
-const state = {
-    cart: [],
-    motifs: [],
-    products: []
-}
-
-const getters = {
-    cart: state => {
-        return state.cart
-    },
-    motifs: state => {
-        return state.motifs
-    },
-    products: state => {
-        return state.products
-    }
-}
-
-const mutations = {
-    ADD_TO_CART(state, payload) {
-        state.cart.push(payload)
-    },
-    SET_CURRENT_ITEM(state, payload) {
-        state.cart = state.cart.map(cartItem => {
-            cartItem.currentEdit = false
-            if(cartItem.cartItemId === payload) cartItem.currentEdit = true
-            return cartItem
-        })
-    },
-    SET_CURRENT_PRODUCT(state, payload) {
-        state.cart = state.cart.map(cartItem => {
-            if(cartItem.currentEdit) cartItem.products = cartItem.products.map(product => {
-                product.currentEdit = false
-                if(product.id === payload) product.currentEdit = true
-                return product
-            })
-            return cartItem
-        })
-    },
-    UPDATE_CART_ITEM(state, payload) {
-        let idx = state.cart.findIndex(cartItem => cartItem.currentEdit == true)
-        state.cart[idx] = payload
-    },
-    REMOVE_FROM_CART(state, payload) {
-        let idx = state.cart.findIndex(cartItem => cartItem.cartItemId === payload.cartItemId)
-        state.cart.splice(idx, 1)
-    },
-    RESET_CART(state) {
-        state.cart = []
-    },
-    SET_MOTIFS(state, motifs) {
-        state.motifs = motifs
-    },
-    SET_PRODUCTS(state, products) {
-        state.products = products
-    }
-}
-
-export const actions = {
+const actions = {
     addToCart({commit, state}, item) {
         let commitItem = baseItem
         
@@ -103,24 +37,6 @@ export const actions = {
     removeFromCart({commit}, item) {
         commit('REMOVE_FROM_CART', item)
     },
-    fetchMotifs({commit}) {
-        axios
-            .get('/api/motifs')
-            .then(res => {
-                //commit('SET_MOTIFS', res.data)
-                commit('SET_MOTIFS', exampleMotifs.data)
-            })
-            .catch(err => console.log(err))
-    },
-    fetchProducts({commit}) {
-        axios
-            .get('/api/products')
-            .then(res => {
-                //commit('SET_PRODUCTS', res.data.data)
-                commit('SET_PRODUCTS', exampleProducts.data)
-            })
-            .catch(err => console.log(err))
-    },
     setTextileColor({commit}, item) {
         let commitItem = cart.find(cartItem => cartItem.currentEdit === true)
         let productIdx = commitItem.products.findIndex(product => product.currentEdit === true)
@@ -155,13 +71,7 @@ export const actions = {
     
         commit('UPDATE_CART_ITEM', commitItem)
     }
+
 }
 
-const store = {
-    state,
-    getters,
-    mutations,
-    actions
-}
-
-export default new Vuex.Store(store);
+export default actions
