+++ resources/js/Store/modules/cart/getters.js
@@ -1,0 +1,7 @@
+const getters = {
+    cart: state => {
+        return state.cart
+    },
+}
+
+export default getters
