+++ resources/js/Store/modules/cart/index.js
@@ -1,0 +1,15 @@
+import getters from './getters'
+import actions from './actions'
+import mutations from './mutations'
+
+const state = () => ({
+    cart: []
+})
+
+export default {
+    namespaced: true,
+    state,
+    getters,
+    actions,
+    mutations
+}
