+++ resources/js/Store/modules/shop/getters.js
@@ -1,0 +1,10 @@
+const getters = {
+    motifs: state => {
+        return state.motifs
+    },
+    products: state => {
+        return state.products
+    }
+}
+
+export default getters
