+++ resources/js/Store/modules/shop/index.js
@@ -1,0 +1,16 @@
+import getters from './getters'
+import actions from './actions'
+import mutations from './mutations'
+
+const state = () => ({
+    products: [],
+    motifs: []
+})
+
+export default {
+    namespaced: true,
+    state,
+    getters,
+    actions,
+    mutations
+}
