+++ resources/js/Store/modules/shop/mutations.js
@@ -1,0 +1,10 @@
+const mutations = {
+    SET_MOTIFS(state, motifs) {
+        state.motifs = motifs
+    },
+    SET_PRODUCTS(state, products) {
+        state.products = products
+    }
+}
+
+export default mutations
